Refactor: 관리자 layout 제거
