+++ frontend/src/components/feature/layout/admin/index.ts
@@ -1,1 +1,1 @@
-export { AdminLayout } from './AdminLayout'
+export { AdminSidebar } from './Sidebar'
