--- frontend/src/components/feature/layout/admin/sidebar/index.ts
@@ -1,1 +1,0 @@
-export { AdminSidebar } from './Sidebar'
