+++ frontend/src/components/feature/layout/main/MainLayout.tsx
@@ -1,12 +1,18 @@
+import { ReactNode } from 'react'
+
 import { Footer } from './footer'
 import { Header } from './header'
 
-export const MainLayout = ({ children }: { children: React.ReactNode }) => {
+interface MainLayoutProps {
+  children: ReactNode
+}
+
+export const MainLayout = ({ children }: MainLayoutProps) => {
   return (
     <main className="flex min-h-screen w-full flex-col">
       <Header />
-      <div className="mt-16 flex w-full flex-1 justify-center px-5 pt-10 sm:px-20">
-        <div className="max-w-screen-xl flex-1">{children}</div>
+      <div className="flex w-full flex-1 justify-center">
+        <div className="mt-16 max-w-screen-2xl flex-1">{children}</div>
       </div>
       <Footer />
     </main>
