+++ frontend/src/routes/Router.tsx
@@ -44,11 +44,12 @@ export const Router = () => {
           <Route path="login" element={<LoginPage />} />
           <Route path="signup" element={<SignupPage />} />
         </Route>
+
         <Route path="/admin" element={<AdminRoute />}>
           <Route path="member" element={<AdminMemberPage />} />
           <Route path="semester" element={<AdminSemesterPage />} />
         </Route>
-        <Route path="/" element={<MainPage />} />
+
         <Route element={<MainRoute />}>
           <Route path="/activity" element={<ActivityRoute />}>
             <Route index element={<SemesterRedirectPage />} />
@@ -93,6 +94,8 @@ export const Router = () => {
             }
           />
         </Route>
+
+        <Route path="/" element={<MainPage />} />
         <Route path="/recruit" element={<RecruitPage />} />
         <Route path="/*" element={<NotFoundPage />} />
       </Routes>
