+++ frontend/src/routes/custom-route/main/MainRoute.tsx
@@ -5,7 +5,9 @@ import { MainLayout } from '@/components/feature/layout/main'
 export const MainRoute = () => {
   return (
     <MainLayout>
-      <Outlet />
+      <div className="flex w-full flex-1 justify-center px-5 pt-10 sm:px-20">
+        <Outlet />
+      </div>
     </MainLayout>
   )
 }
